02-transformacao-dados/src/requirements.txt
tabula-py==2.4.0
pandas==1.3.4